# docs/openapi/driver.json
{
  "components": {
    "schemas": {
      "AdminDriverResponse": {
        "properties": {
          "accountStatus": {
            "type": "string"
          },
          "activeDeliveryId": {
            "type": "string"
          },
          "activeRouteId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currentLat": {
            "type": "number"
          },
          "currentLng": {
            "type": "number"
          },
          "delivered": {
            "format": "int32",
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "failed": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "invitationExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "invitedByName": {
            "type": "string"
          },
          "lastInvitedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLocationAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "onlineStatus": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "registered": {
            "type": "boolean"
          },
          "suspendedReason": {
            "type": "string"
          },
          "totalDeliveries": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DriverProfileResponse": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "currentLat": {
            "type": "number"
          },
          "currentLng": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "lastLocationAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "onboardingStatus": {
            "type": "string"
          },
          "onlineStatus": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "HistoryItem": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "deliveryId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "HistoryResponse": {
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "LocationRequest": {
        "properties": {
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "type": "object"
      },
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "SetStatusRequest": {
        "properties": {
          "isRegistered": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "StatsResponse": {
        "properties": {
          "cancelled": {
            "format": "int32",
            "type": "integer"
          },
          "delivered": {
            "format": "int32",
            "type": "integer"
          },
          "failed": {
            "format": "int32",
            "type": "integer"
          },
          "totalDeliveries": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "bearerFormat": "JWT",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "type": "http"
      },
      "Internal Secret": {
        "in": "header",
        "name": "X-Internal-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API documentation for the ASM Driver Service",
    "title": "Driver Service API",
    "version": "v1.0.0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/admin/drivers/{id}/status": {
      "patch": {
        "operationId": "setStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDriverResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Activate or deactivate driver (PENDING_SETUP is rejected)",
        "tags": [
          "Admin Drivers"
        ]
      }
    },
...
    "/api/v1/driver/availability": {
      "patch": {
        "description": "Sets the driver's online status: ONLINE, ON_BREAK or OFFLINE.",
        "operationId": "updateAvailability",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Set my availability",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/fcm-token": {
      "delete": {
        "description": "Removes the stored FCM token (e.g. on logout).",
        "operationId": "clearFcmToken",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Clear my push token",
        "tags": [
          "Driver Operations"
        ]
      },
      "put": {
        "description": "Stores the device's FCM token for push notifications.",
        "operationId": "updateFcmToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Register my push token",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/history": {
      "get": {
        "description": "Paginated history of the driver's past deliveries.",
        "operationId": "getHistory",
        "parameters": [
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get my delivery history",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/location": {
      "post": {
        "description": "Pushes the driver's live GPS position (used for tracking).",
        "operationId": "updateLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Report my current location",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/me/photo": {
      "post": {
        "description": "Uploads/replaces the driver's own photo (multipart field 'file'); re-encoded and stored server-side.",
        "operationId": "uploadPhoto",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upload my profile photo",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/profile": {
      "get": {
        "description": "Returns the signed-in driver's own profile.",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverProfileResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get my driver profile",
        "tags": [
          "Driver Operations"
        ]
      }
    },
    "/api/v1/driver/stats": {
      "get": {
        "description": "Returns the driver's own performance counters.",
        "operationId": "getStats",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get my delivery stats",
        "tags": [
          "Driver Operations"
        ]
      }
    },
...
  },
  "security": [
    {
      "Bearer Authentication": [],
      "Internal Secret": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://driver-service:8086"
    }
  ],
...
}